Fix CSV reader when `sep=None` (#1455)
* Resolves #1452

Change the way we check for sep=False. Previously, pandas allowed
sep=False for tab delimeted text files. Because we explicitly look for
a sep character we converted sep=False to sep="\t". We checked
with a line if not sep which also evaluated to True when sep is
None. sep=None has a different behavior as outlined in #1452.