FEAT-6225: Allow set_index to take an object to be handled by the backend (a backend index, etc)
As discussed in #1780, the set_index API assumes that we either have a set of column names/objects and that the set of columns passed in matches the set of known columns. We then call set_index_from_columns to let the backend handle the operation. A problem occurs when we need to call set_index with a backend type which the modin layer does not know about:

df.set_index(MyBackendIndex)
MyBackendIndex is not a normal Pandas type, nor a type which modin knows about, nor an actual column in the dataframe. In this case we want the backend to be responsible for as much of the handling here as possible. We also want to retain the nice error handling already present in set_index to get for missing columns.

One solution I have is to get to see if keys is non-"primitive" and when it is considered missing from the columns we raising an error and throw it over the wall to the query_compiler. Another method might be to agree on an attribute or method set on each backend which can tell the dataframe set_index method to circumvent the column checks.

Signed-off-by: John Kew <jkew@ponder.io>