FEAT-#6284: Do not convert HDK query execution result to arrow.
Now, the execution result is not converted to arrow but is stored
in partitions as the ExecutionResultTable object. This object
represents a table in the HDK database and provides some basic information
about the table - name, column_names, shape. If an arrow table is
imported to HDK, a corresponding ImportedTable object is stored in
partitions. All the subsequent operations are performed, using the
*Table objects, without conversion to/from arrow.

Signed-off-by: Andrey Pavlenko <andrey.a.pavlenko@gmail.com>