FIX-#6778: Read parquet files without file extensions using fastparquet
In supporting fastparquet, modin takes the paths provided, globs them,
and filters them to only look at files with the .parq or .parquet
extension. This commit adds support so that if the path supplied is
explicitly a file, it will be included.

This commit also fixes the formatting issues for the `black` linter.

Signed-off by: Ari Brown <ari@aribrown.com>