FEAT-#7448: Allow QueryCompilerCaster to apply cost-optimization on automatic casting (#7464)
This CL introduces the ability to symmetrically cast to different
query compilers based on a cost estimation. Each query compiler
implements a new function which can return a cost estimate of data
migration from that specific compiler's perspective. On each operation
involving mixed types we calculate the aggregate cost of migration
across all query compilers and pick the query compiler class with the
least cost.

The implementation of the cost-estimate, from each query compiler,
can be anything, but it is constrained to a set of values between 0-1000.

Max Cost Not Implemented
This implementation still can face an issue where a workload is moved
to the wrong engine, because there is no threshold on the cost of migration
implemented. This can be implemented the future.

---------

Co-authored-by: Jonathan Shi <149419494+sfc-gh-joshi@users.noreply.github.com>
Co-authored-by: Mahesh Vashishtha <mahesh.vashishtha@snowflake.com>