REFACTOR-#7470: Combine backend casting and extension code at the API layer. (#7485)
Prior to this commit, we had both the QueryCompilerCaster, which cast query compilers from different backends to common backend, and the extension system, which selected an extension method that matched a backend. This commit moves the extension code into the QueryCompilerCaster, which now selects the correct backend, casts function inputs to the correct backend, finds the correct extension function if it exists, and then applies the correct function to the inputs.

Signed-off-by: sfc-gh-mvashishtha <mahesh.vashishtha@snowflake.com>