REFACTOR-#7534: Provide internal, overridable method for "estimated row count"
Existing query compilers can call get_axis_len(axis=0) to get a row count, but
this may be expensive on lazily evaluated systems. Instead we add a new function
which can be overridden on a per-query compiler basis for the estimated row count.

The default implementation of this is to simply call get_axis_len(axis=0) but
sub-classes may have cheaper ways of estimating the length of the dataset.