REFACTOR-#7534: Provide internal, overridable method for max_shape (#7535)
Existing query compilers can call get_axis_len(axis=0) to get a row
count, but this may be expensive on lazily evaluated systems. Instead we
add a new function which can be overridden on a per-query compiler basis
for the max row count, which can be sometimes calculated more cheaply.

The default implementation of this is to simply call get_axis_len but
sub-classes may have cheaper ways of estimating the length of the
dataset.

<!--
Thank you for your contribution!
Please review the contributing docs:
https://modin.readthedocs.io/en/latest/development/contributing.html
if you have questions about contributing.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

- [x] first commit message and PR title follow format outlined
[here](https://modin.readthedocs.io/en/latest/development/contributing.html#commit-message-formatting)
> **_NOTE:_** If you edit the PR title to match this format, you need to
add another commit (even if it's empty) or amend your last commit for
the CI job that checks the PR title to pick up the new PR title.
- [ ] passes `flake8 modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [ ] passes `black --check modin/ asv_bench/benchmarks
scripts/doc_checker.py`
- [x] signed commit with `git commit -s` <!-- you can amend your commit
with a signature via `git commit -amend -s` -->
- [ ] Resolves #7534 <!-- issue must be created for each patch -->
- [ ] tests added and passing
- [x] module layout described at `docs/development/architecture.rst` is
up-to-date <!-- if you have added, renamed or removed files or
directories please update the documentation accordingly -->