FIX-#7528: Dispatch module-level extensions to the correct backend (#7529)
In this PR, the extensions system tracks objects that were re-exported
from pandas, and either wraps them in
`wrap_free_function_in_argument_caster` if it is a function, or removes
them from the namespace and dispatches via `__getattr__` if it is a
class. Technically the `__getattr__` approach can be used for all
objects, but this remains more consistent with how we handle dispatch
for top-level functions that modin itself defines.

Signed-off-by: Jonathan Shi <jonathan.shi@snowflake.com>