FEAT-#7542: Support post-operation backend switch for groupby. (#7545)
Make DataFrameGroupBy and SeriesGroupBy inherit from QueryCompilerCaster
so that they can do post-operation automatic backend switch.

Defer pre-operation switch (#7544) until we implement `set_backend()`
and `_copy_into()`, which are more complicated for the groupby objects.

I have validated that the automatic switching post-groupby that we want
for snowpark-python can work with this commit.

Resolves #7542

---------

Signed-off-by: sfc-gh-mvashishtha <mahesh.vashishtha@snowflake.com>