TEST-#7643: Fix residual failures from pandas 2.3 (#7644)
This PR fixes 2 failures I missed while adding support for pandas 2.3:
- `DataFrame.insert` apparently is now typed with `pandas-stubs==2.3`,
but was not typed before. Since pandas 2.3 is compatible with Python 3.9
all CI runs should pick up this change, but if it continues to be an
issue I'll look into a workaround for seeing if it's possible to
conditionally add the `type: ignore` directive.
- `np.cumproduct` was
[deprecated](https://numpy.org/doc/1.26/release/1.25.0-notes.html#deprecations)
some time ago, and removed in numpy>2. We've supported numpy>2 for a
while, so it's unclear why this issue only just triggered.

Signed-off-by: Jonathan Shi <jonathan.shi@snowflake.com>