FIX-#7653: Respect `AutoSwitchBackend` for `DataFrame.T`/`Series.T~ (#7654)
Currently, `DataFrame.T` is declared like so:

```python

    def transpose(self, copy=False, *args) -> DataFrame:  # noqa: PR01, RT01, D200
        """
        Transpose index and columns.
        """
        # FIXME: Judging by pandas docs `*args` serves only compatibility purpose
        # and does not affect the result, we shouldn't pass it to the query compiler.
        return self.__constructor__(
            query_compiler=self._query_compiler.transpose(*args)
        )

    T: DataFrame = property(transpose)
```

This breaks automatic backend switching in a very subtle way. Because we
use the `QueryCompilerCaster` ABC to replace all methods with a wrapper
that picks the correct backend, the `transpose` referenced in
`property(transpose)` is actually the original, unwrapped method, and
therefore never uses the switching code path. This PR replaces the
definition of `T` with
```python
    @property
    def T(self):
        return self.transpose()
```
which defers lookup of `transpose` to runtime, and therefore uses the
caster wrapper correctly.

I initially attempted to add special logic to wrap `fget`/`fset`/`fdel`
or property objects in `QueryCompilerCaster`, but this became
complicated and also didn't work for some reason--somehow, even though I
was able to wrap `fget`, the wrapped function was never called. The
approach in this PR is simpler, and runs a lower risk of breaking
behavior or accidentally causing thrashing for other commonly-accessed
properties like `columns` and `index`.

---------

Signed-off-by: Jonathan Shi <jonathan.shi@snowflake.com>