FIX-#7664: Add typing_extensions dependency (#7665)
`typing_extensions` (used to get the `Self` type) is included as a
tertiary dependency of Ray for Python 3.12 and below, but it is no
longer included for 3.13. This PR makes it a direct dependency of modin.

The [conda-forge](https://github.com/conda-forge/modin-feedstock) recipe
(which has been broken for a while) will need to be updated.

Signed-off-by: Jonathan Shi <jonathan.shi@snowflake.com>