FIX-#7669: Respect eval(inplace=False). (#7670)
Prior to this commit, calling `DataFrame.eval(inplace=False,
engine='python')` with an assignment would incorrectly update the input
dataframe. The root cause was calling `copy(deep=None)` instead of
`copy(deep=True)`. We need to insert new columns to a deep copy of the
input dataframe. `deep=None` is not part of the `copy` API at all, so we
don't need to fix `copy()` behavior. It happens that `deep=None` does a
deep copy in pandas, which is where we got the `deep=None` code from.

Resolves #7669

Signed-off-by: sfc-gh-mvashishtha <mahesh.vashishtha@snowflake.com>