FEAT-#7676: in-place casting between DataFrame engines (#7666)
Addresses a performance issue with hybrid execution where a sequence of
merge operations using the same DataFrames could result in transfer
thrashing. In this PR we have the arguments of the operation move
in-place so that subsequent operations all stay on the same backend.
This behavior can be turned off by setting the `BackendMergeCastInPlace`
variable to false.

<!-- Please give a short brief about these changes. -->

- [x] first commit message and PR title follow format outlined
[here](https://modin.readthedocs.io/en/latest/development/contributing.html#commit-message-formatting)
> **_NOTE:_** If you edit the PR title to match this format, you need to
add another commit (even if it's empty) or amend your last commit for
the CI job that checks the PR title to pick up the new PR title.
- [x] passes `flake8 modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] passes `black --check modin/ asv_bench/benchmarks
scripts/doc_checker.py`
- [x] signed commit with `git commit -s` <!-- you can amend your commit
with a signature via `git commit -amend -s` -->
- [x] Resolves #7676 <!-- issue must be created for each patch -->
- [x] tests added and passing
- [x] module layout described at `docs/development/architecture.rst` is
up-to-date <!-- if you have added, renamed or removed files or
directories please update the documentation accordingly -->

---------

Co-authored-by: Jonathan Shi <149419494+sfc-gh-joshi@users.noreply.github.com>
Co-authored-by: Mahesh Vashishtha <mahesh.vashishtha@snowflake.com>